코멘트 라우터,컨트롤러 ts 변환 (상품 댓글은 아직 보류)
