+++ src/structs/userStruct.ts
@@ -3,11 +3,11 @@ import isEmail from 'is-email';
 import type { Infer } from 'superstruct';
 import { IdParam, PaginationQuery, OrderQuery } from './commonStruct.js';
 
-const emailValidator = (value: unknown): boolean => {
+const emailValidator = (value: unknown): value is string => {
   return typeof value === 'string' && isEmail(value);
 };
 export const CreateUser = s.object({
-  email: s.define('Email', emailValidator), //s.define(<> ,...) <> =검사에서 실패했을 때 에러메시지에 사용
+  email: s.refine(s.string(), 'Email', emailValidator), //s.define(<> ,...) <> =검사에서 실패했을 때 에러메시지에 사용
   password: s.size(s.string(), 8, 20),
   nickname: s.size(s.string(), 2, 10),
   name: s.size(s.string(), 1, 10),
@@ -19,7 +19,7 @@ export const PatchUser = s.partial(s.omit(CreateUser, ['password']));
 export type PatchUserType = Infer<typeof PatchUser>;
 
 export const LoginUser = s.object({
-  email: s.define('Email', emailValidator),
+  email: s.refine(s.string(), 'Email', emailValidator),
   password: s.size(s.string(), 8, 20),
 });
 export type LoginUserType = Infer<typeof LoginUser>;
